Add redirection logic based on auth
